+++ R/query-set-op.R
@@ -1,11 +1,12 @@
 #' @export
 #' @rdname sql_build
-set_op_query <- function(x, y, type = type) {
+set_op_query <- function(x, y, type = type, all = FALSE) {
   structure(
     list(
       x = x,
       y = y,
-      type = type
+      type = type,
+      all = all
     ),
     class = c("set_op_query", "query")
   )
@@ -27,5 +28,12 @@ sql_render.set_op_query <- function(query, con = NULL, ..., subquery = FALSE) {
   from_x <- sql_render(query$x, con, ..., subquery = FALSE)
   from_y <- sql_render(query$y, con, ..., subquery = FALSE)
 
-  dbplyr_query_set_op(con, from_x, from_y, method = query$type)
+  if (dbplyr_edition(con) >= 2) {
+    sql_query_set_op(con, from_x, from_y, method = query$type, all = query$all)
+  } else {
+    if (isTRUE(query$all)) {
+      abort("`all` argument not supported by this backend")
+    }
+    dbplyr_query_set_op(con, from_x, from_y, method = query$type)
+  }
 }